adding day 14 work trying to deploy backend and making some file starcture and frontend work
