+++ BackEnd/config/db.js
@@ -3,7 +3,6 @@ const mongoose = require("mongoose")
 mongoose.set('strictQuery', false)
 require('dotenv').config();
 
-
 connect = mongoose.connect(process.env.MONGOOSE_URL)
 
 module.exports = { connect }
