+++ FrontEnd/stylecapsule/src/components/HomePage/Footer/Component/02-FooterTextBar/FooterTextBar.jsx
