+++ FrontEnd/stylecapsule/src/components/HomePage/Home/Home.jsx
@@ -1,20 +1,32 @@
-import { Button, Heading } from "@chakra-ui/react";
 import React from "react";
-import { CommentBox } from "./Pages/Comment/CommentBox";
-import Carousel from "./Pages/FirstBox/FirstBox"
-import { Flowers_Check } from "./Pages/FlowersCheck/Flowers_Check";
-import { Help_Image_box } from "./Pages/HelpIMageBox/Help_Image_box";
-import { New_Arrivals } from "./Pages/New Arrivals/New_Arrivals";
-import { WelcomeBox } from "./Pages/WelcomeBox/WelcomeBox";
+import { ImageBox } from "./Pages/01-ImageBox.Home";
+import { WelcomeTextBox } from "./Pages/02-WelcomeTextBox.Home.";
+import { ProductDetailsTextBox } from "./Pages/03-ProductDetailsTextBox.Home";
+import { CommentSliderBox } from "./Pages/04-CommentSliderBox.Home";
+import { ThankYouBox } from "./Pages/05-ThankYouBox.Home";
+
+
 export const Home = () => {
   return (
     <div className="HomePage">
-      <Carousel  />
-      <WelcomeBox />
-      {/* <New_Arrivals /> */}
-      <Flowers_Check />
-      <CommentBox />
-      <Help_Image_box />
+
+    {/* image Box first of flower   */}
+      <ImageBox  /> 
+
+    {/* welcome Box text to introduce about website */}
+      <WelcomeTextBox />
+
+      {/* <New_Arrivals />  trying to add product slider */}
+
+
+      {/* some details about website   */}
+      <ProductDetailsTextBox />
+
+       {/* some comment about website in slider  */}
+      <CommentSliderBox />
+
+       {/* last box to make user good and some help text written */}
+      <ThankYouBox />
     </div>
   );
 };
