--- FrontEnd/stylecapsule/src/components/HomePage/Home/Pages/Comment/CommentBox.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { SliderComment } from './SliderComment/SliderComment.js'
-
-export const CommentBox = () => {
-    return (
-        <div>
-            <SliderComment />
-        </div>
-    )
-}
