+++ FrontEnd/stylecapsule/src/components/page/SingleProductAddToCartBox/ProductNameBox/NameBox.jsx
