+++ FrontEnd/stylecapsule/src/components/page/Valentine/ValentinePage.jsx
@@ -1,13 +1,13 @@
-import { Box, Button, Flex, SimpleGrid } from "@chakra-ui/react";
+import { Box, Flex, SimpleGrid } from "@chakra-ui/react";
 import { Spinner } from "@chakra-ui/react";
 
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { FilterBox } from "./FilterBox/FilterBox";
-import { SingleProductBox } from "../SingleProductBox/SingleProductBox";
+import { SingleProductAddToCartBox } from "../SingleProductAddToCartBox/SingleProductAddToCartBox";
 import { PaginationBox } from "./Pagination/PaginationBox";
-import { useSelector } from "react-redux";
-import { Set_Single_Page_data } from "../../Redux/products/Prodaction";
+// import { useSelector } from "react-redux";
+// import { Set_Single_Page_data } from "../../Redux/products/Prodaction";
 
 export const ValentinePage = () => {
   const [Loading, setLoading] = useState(false);
@@ -17,7 +17,7 @@ export const ValentinePage = () => {
 
 
 
-  const getData = async () => {
+  const GetData = async (page) => {
     try {
       return await axios.get(
         `http://localhost:4000/valentine_Day/?page=${page}&limit=8`
@@ -28,7 +28,7 @@ export const ValentinePage = () => {
   };
   useEffect(() => {
     setLoading(true);
-    getData(page)
+    GetData(page)
       .then((res) => {
         setList(res.data.valentine);
         setTotal(res.data.totalPages);
@@ -37,7 +37,7 @@ export const ValentinePage = () => {
       .catch((err) => console.log(err));
   }, [page]);
 
-  console.log(list);
+
   return (
     <>
       <Flex  gap="8"  display={["grid", "grid", "grid", "flex"]} w="95%"  m="auto" mb="2rem" mt="2rem">
@@ -66,7 +66,7 @@ export const ValentinePage = () => {
           ) : (
             <SimpleGrid columns={[1, 2, 2, 4]} spacing={10} display="grid">
               {list.map((product, i) => (
-                <SingleProductBox product={product} i={i} />
+                <SingleProductAddToCartBox product={product} i={i} />
               ))}
             </SimpleGrid>
           )}
